Some refactoring
Trying to separate functions out of the GroceryView and into a GroceryViewModel to perform these functions
