+++ Groceries/App/Assets.xcassets/Contents.json
