+++ Groceries/App/Assets.xcassets/Groceries/apple.imageset/Contents.json
