+++ Groceries/App/Assets.xcassets/Groceries/apple.imageset/apple.png
