+++ Groceries/App/Assets.xcassets/Groceries/cherry.imageset/Contents.json
