+++ Groceries/App/Assets.xcassets/Groceries/cherry.imageset/cherry.png
