+++ Groceries/App/Assets.xcassets/Groceries/fruitloops.imageset/fbc211_clipped_rev_1.png
