+++ Groceries/App/Assets.xcassets/Groceries/gooseberry.imageset/Contents.json
