+++ Groceries/App/Assets.xcassets/Groceries/lemon.imageset/Contents.json
