+++ Groceries/App/Assets.xcassets/Groceries/lemon.imageset/lemon.png
