+++ Groceries/App/Assets.xcassets/Groceries/lime.imageset/Contents.json
