+++ Groceries/App/Assets.xcassets/Groceries/mango.imageset/Contents.json
