+++ Groceries/App/Assets.xcassets/Groceries/strawberry.imageset/Contents.json
