+++ Groceries/App/Assets.xcassets/Groceries/strawberry.imageset/strawberry.png
