+++ Groceries/App/Assets.xcassets/Groceries/watermelon.imageset/Contents.json
