+++ Groceries/App/Assets.xcassets/Groceries/whitebread.imageset/fbb170_clipped_rev_1.png
